refactor(vestad): delete control_ws and the agents_ws_message fixture, relocate invalidate_service_handler
